fix: change quote_increment type from double to decimal(32,16) in g_product table

+++ ddl.sql
@@ -95,7 +95,7 @@ CREATE TABLE `g_product` (
   `base_max_size` decimal(32,16) NOT NULL,
   `base_scale` int(11) NOT NULL,
   `quote_scale` int(11) NOT NULL,
-  `quote_increment` double NOT NULL,
+  `quote_increment` decimal(32,16) NOT NULL,
   `quote_min_size` decimal(32,16) NOT NULL,
   `quote_max_size` decimal(32,16) NOT NULL,
   PRIMARY KEY (`id`)
